refactor: extract useStats hook into a separate file for better modularity and fast refresh compatibility
